Refactor common reconnection algorithm into base class
